Fix review comments for macro imports
